+++ .gitignore
@@ -43,3 +43,7 @@ app.*.map.json
 !/packages/flutter_tools/test/data/dart_dependencies_test/**/.packages
 
 pubspec.lock
+
+android/app/google-services.json
+ios/GoogleService-Info.plist
+ios/Podfile.lock
