+++ lib/main.dart
@@ -1,39 +1,11 @@
 import 'package:dumka/utils/const.dart';
-import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 
 import 'screens/splash_screen.dart';
 
 void main() {
-  runApp(App());
-}
-
-class App extends StatelessWidget {
-  // Create the initialization Future outside of `build`:
-  final Future<FirebaseApp> _initialization = Firebase.initializeApp();
-
-  @override
-  Widget build(BuildContext context) {
-    return FutureBuilder(
-      // Initialize FlutterFire:
-      future: _initialization,
-      builder: (context, snapshot) {
-        // Check for errors
-        // if (snapshot.hasError) {
-        //   return SomethingWentWrong();
-        // }
-
-        // Once complete, show your application
-        // if (snapshot.connectionState == ConnectionState.done) {
-          return MyApp();
-        // }
-
-        // Otherwise, show something whilst waiting for initialization to complete
-        // return Loading();
-      },
-    );
-  }
+  runApp(MyApp());
 }
 
 class MyApp extends StatelessWidget {
@@ -55,7 +27,7 @@ class MyApp extends StatelessWidget {
         theme: ThemeData(
           visualDensity: VisualDensity.adaptivePlatformDensity,
         ),
-        home: SplashScreen());
+        home: InitScreen());
   }
 }
