+++ lib/screens/splash_screen.dart
@@ -1,12 +1,40 @@
 import 'package:dumka/screens/proposals_screen.dart';
 import 'package:dumka/utils/const.dart';
+import 'package:firebase_core/firebase_core.dart';
+import 'package:firebase_crashlytics/firebase_crashlytics.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 
-class SplashScreen extends StatelessWidget {
+
+class InitScreen extends StatelessWidget {
+  // Create the initialization Future outside of `build`:
+  final Future<FirebaseApp> _initialization = Firebase.initializeApp();
+
   @override
   Widget build(BuildContext context) {
+    return FutureBuilder(
+      // Initialize FlutterFire:
+      future: _initialization,
+      builder: (context, snapshot) {
+        // Check for errors
+        if (snapshot.hasError) {
+          pushToProposalsScreens(context);
+        }
+
+        // Once complete, show your application
+        if (snapshot.connectionState == ConnectionState.done) {
+          FirebaseCrashlytics.instance.setCrashlyticsCollectionEnabled(true);
+          pushToProposalsScreens(context);
+        }
+
+        // Otherwise, show something whilst waiting for initialization to complete
+        return SplashScreen();
+      },
+    );
+  }
+
+  void pushToProposalsScreens(BuildContext context) {
     Future.delayed(const Duration(seconds: 1), () {
       Navigator.pushReplacement(
           context,
@@ -14,7 +42,14 @@ class SplashScreen extends StatelessWidget {
             builder: (context) => ProposalsScreen(),
           ));
     });
+  }
 
+}
+
+class SplashScreen extends StatelessWidget {
+
+  @override
+  Widget build(BuildContext context) {
     return Scaffold(
       backgroundColor: Colors.white,
       body: Stack(
